apps/api/requirements.txt
# FastAPI and server
fastapi==0.110.1
uvicorn[standard]==0.29.0
python-multipart==0.0.9
httpx==0.27.0

# Database
sqlalchemy[asyncio]==2.0.29
asyncpg==0.29.0
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
supabase==2.4.0

# Validation
pydantic==2.7.0
pydantic-settings==2.2.1
email-validator==2.1.1

# ML/AI
scikit-learn==1.4.2
joblib==1.4.0
numpy==1.26.4
pandas==2.2.2

# Background tasks
apscheduler==3.10.4
redis==5.0.3

# Utilities
python-dotenv==1.0.1
python-dateutil==2.9.0

# Testing
pytest==8.1.1
pytest-asyncio==0.23.6
pytest-cov==5.0.0

# Linting
ruff==0.3.7
